Consolidate internal ML-DSA files
