--- crypto/internal/mldsa87/ntt.go
@@ -1,11 +1,0 @@
-package mldsa87
-
-import "github.com/theQRL/go-qrllib/crypto/internal/lattice"
-
-func ntt(a *[N]int32) {
-	lattice.NTT(a)
-}
-
-func invNTTToMont(a *[N]int32) {
-	lattice.InvNTTToMont(a)
-}
