+++ crypto/internal/mldsa87/structured_fuzz_test.go
@@ -189,3 +189,21 @@ func FuzzPrivateKeyFromSeedSignVerify(f *testing.F) {
 		}
 	})
 }
+
+// FuzzPrivateKeyVerify tests that Verify handles arbitrary input without panicking.
+func FuzzPrivateKeyVerify(f *testing.F) {
+	// Add seed corpus with various sizes.
+	f.Add(make([]byte, 0), make([]byte, 0), make([]byte, CRYPTO_BYTES), make([]byte, CRYPTO_PUBLIC_KEY_BYTES))
+	f.Add(make([]byte, 10), make([]byte, 32), make([]byte, CRYPTO_BYTES), make([]byte, CRYPTO_PUBLIC_KEY_BYTES))
+	f.Add(make([]byte, 255), make([]byte, 1000), make([]byte, CRYPTO_BYTES), make([]byte, CRYPTO_PUBLIC_KEY_BYTES))
+
+	f.Fuzz(func(t *testing.T, ctx, message, sigBytes, pkBytes []byte) {
+		var sig [CRYPTO_BYTES]uint8
+		var pk [CRYPTO_PUBLIC_KEY_BYTES]uint8
+
+		copy(sig[:], sigBytes)
+		copy(pk[:], pkBytes)
+
+		_ = verifyForTest(ctx, message, sig, &pk)
+	})
+}
